Shuffle around memory provider
