+++ vortex-datafusion/src/memory/mod.rs
@@ -1,0 +1,11 @@
+mod exec;
+mod plans;
+mod provider;
+mod statistics;
+mod stream;
+
+pub use exec::*;
+pub use plans::*;
+pub use provider::*;
+pub use statistics::*;
+pub use stream::*;
