+++ vortex-datafusion/src/memory/statistics.rs
